Patch largely from Josh from bug #44539 - Support for area references in formulas of rows >= 32768
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@634630 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/record/formula/AreaI.java
@@ -24,22 +24,22 @@ public interface AreaI {
     /**
      * @return the first row in the area
      */
-    public short getFirstRow();
+    public int getFirstRow();
 
     /**
      * @return last row in the range (x2 in x1,y1-x2,y2)
      */
-    public short getLastRow();
+    public int getLastRow();
     
     /**
      * @return the first column number in the area.
      */
-    public short getFirstColumn();
+    public int getFirstColumn();
     
     /**
      * @return lastcolumn in the area
      */
-    public short getLastColumn();
+    public int getLastColumn();
     
     /**
      * @return isrelative first column to relative or not
